+++ crates/python/src/qvm/api.rs
@@ -77,7 +77,7 @@ impl PyAddressRequest {
     }
 
     #[staticmethod]
-    pub fn from_indices(indices: Vec<usize>) -> Self {
+    pub fn from_indices(indices: Vec<u32>) -> Self {
         Self(AddressRequest::Indices(indices))
     }
 }
